compile-time enforcement of printf strings only:uaefi

+++ firmware/console/eficonsole.cpp
@@ -53,7 +53,7 @@ static void printUid() {
 
 static void sayHello() {
 	efiPrintf(PROTOCOL_HELLO_PREFIX " rusEFI LLC (c) 2012-2024. All rights reserved.");
-	efiPrintf(PROTOCOL_HELLO_PREFIX " rusEFI v%d@%d now=%d", getRusEfiVersion(), SIGNATURE_HASH, (int)getTimeNowMs());
+	efiPrintf(PROTOCOL_HELLO_PREFIX " rusEFI v%d@%d now=%d", getRusEfiVersion(), /*do we have a working way to print 64 bit values?!*/(int)SIGNATURE_HASH, (int)getTimeNowMs());
 	efiPrintf(PROTOCOL_HELLO_PREFIX " Chibios Kernel:       %s", CH_KERNEL_VERSION);
 	efiPrintf(PROTOCOL_HELLO_PREFIX " Compiled:     " __DATE__ " - " __TIME__ "");
 	efiPrintf(PROTOCOL_HELLO_PREFIX " COMPILER=%s", __VERSION__);
